requirements.txt
playwright==1.41.0
pandas==2.2.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
